% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listprocessor.R
\name{first}
\alias{first}
\title{first}
\usage{
first(lst)
}
\arguments{
\item{lst}{the list to which the first element will be returned}
}
\value{
the first element
}
\description{
Retrieve the first element from list
}
\details{
Retrieve and return the first element from a list. This lispy
style function uses indexing under the hood but exposes the
process as a function call.
}
\examples{
(first {1 2 3}) => 1
(first {{1 2} {3 4}}) => list(1, 2)
}
\author{
Jay Morgan
}
