% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{not}
\alias{not}
\title{not}
\usage{
not(bool)
}
\arguments{
\item{bool}{The bool to invert \code{TRUE/FALSE}}
}
\value{

}
\description{
invert a bool value, i.e. TRUE => FALSE, FALSE => TRUE
}
\details{
Instead of the traditional inversion of boolean values using
!(bool), here we provide a more lispy style inversion function called 'not'.
This can be applied to boolean functions as a simple function call.
}
\examples{
(not TRUE)
(not FALSE)
(not (== 1 1))
}
\author{
Jay Morgan
}
