% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listprocessor.R
\name{nth}
\alias{nth}
\title{nth}
\usage{
nth(lst, n)
}
\arguments{
\item{lst}{The list to which the nth element will be returned}

\item{n}{the index of the element to retrieve}
}
\value{
the nth element
}
\description{
Return the nth element from a list
}
\details{
Retrieve the nth element of the list via a function call instead of indexing
}
\examples{
(nth {1 2 3} 1) => 1
(nth {{1 2} {3 4}} 2) => list(3, 4)
}
\author{
Jay Morgan
}
