% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listprocessor.R
\name{rest}
\alias{rest}
\title{rest}
\usage{
rest(lst)
}
\arguments{
\item{lst}{the list to get everything but the first index}
}
\value{
the list excluding the first index
}
\description{
Retrieve the rest of a list
}
\details{
Retrieve the rest of the elements from a list type data structure.
}
\examples{
\dontrun{
(rest {1 2 3 4}) => list(2, 3, 4)
(rest {{1 2} {3 4}}) => list(3, 4)
}
}
\author{
Jay Morgan
}
