% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ast.R
\name{slurp_ast}
\alias{slurp_ast}
\title{slurp_ast}
\usage{
slurp_ast(input)
}
\arguments{
\item{input}{as a lisp-style string}
}
\value{
List of lists for each token
}
\description{
Tokenise the user input
}
\details{
From the lisp syntax, tokenise and provide a AST representation using
list of lists.
}
\author{
Jay Morgan
}
