% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{slurp_evaluate_ast}
\alias{slurp_evaluate_ast}
\title{evaluate_ast}
\usage{
slurp_evaluate_ast(ast_list)
}
\arguments{
\item{ast_list}{the results of ast(...)}
}
\value{
the compiled R expression
}
\description{
Evaluate the AST form
}
\details{
Using the result of ast(...), evaluate and compile
the expresions
}
\author{
Jay Morgan
}
