% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{source_slurp}
\alias{source_slurp}
\title{source_slurp}
\usage{
source_slurp(source_code)
}
\arguments{
\item{source_code}{String representation of the source}
}
\description{
Import a SluRp script file
}
\details{
Import a SluRp script file
}
\author{
Jay Morgan
}
